+++ CHANGES.md
@@ -3,6 +3,8 @@
 - Add feedback for alt + f, which toggles personal roboports (not the same as alt + l, which toggles personal logistics).
 - Add a button in the character's equipment GUI to toggle roboports, for those who do not find the key.
 - Risk an internal cleanup refactor to how equipment bonuses are announced
+- Fix dragging stops up and down in the train schedule editor
+- Make the train config panel the first tab for locomotives, since locomotive inventories are rarely accessed
 
 # 0.16.22 (2025-11-20)